LayoutTests/accessibility/ios-simulator/role-switch.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script>
var successfullyParsed = false;
if (window.testRunner)
   testRunner.dumpAsText();
</script>
</head>
<body>

<button role="switch" aria-checked="true" id="switch" class="switch">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This test thats accessibilityValue is correct for role=switch.");

    if (window.accessibilityController) {

        var element = accessibilityController.accessibleElementById("switch");
        shouldBe("element.stringValue", "'AXValue: 1'");

        document.getElementById("switch").ariaChecked = false;
        shouldBe("element.stringValue", "'AXValue: 0'");
    }

    successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>

</body>
</html>
