LayoutTests/accessibility/ios-simulator/role-switch-expected.txt


This test thats accessibilityValue is correct for role=switch.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.stringValue is 'AXValue: 1'
PASS element.stringValue is 'AXValue: 0'
PASS successfullyParsed is true

TEST COMPLETE
